+++ libs/shared/src/lib/layouts/top-bar/index.ts
@@ -1,0 +1,5 @@
+export { BrandComponent } from './brand/brand.component';
+export { EnvironmentBadgeComponent } from './environment-badge/environment-badge.component';
+export { NotificationsComponent } from './notifications/notifications.component';
+export { TopBarComponent } from './top-bar.component';
+export { UserProfileComponent } from './user-profile/user-profile.component';